requirements.txt
dareplane_utils
fire
numpy
pandas
plotly
pyglet
pylsl
pyserial
pyxdf
PyYAML
sounddevice
